+++ pyproject.toml
@@ -27,13 +27,10 @@ dependencies = [
     "altair>=5.3.0",
     "biopython>=1.80",
     "cyclopts>=4.3.0",
-    "defopt==6.4.0",
-    "httpx>=0.27.0",
     "jsonschema",
     "linkml>=1.9.4",
     "natsort>=8.4.0",
     "pandas>=1.5.3",
-    "platformdirs>=4.2.2",
     "primalbedtools>=1.0.0",
     "pydantic>=2.0.0",
     "pyyaml>=6.0.2",
