+++ src/primaschema/__init__.py
@@ -6,8 +6,6 @@ import os
 
 from pathlib import Path
 
-from platformdirs import user_data_dir
-
 
 __version__ = "1.0.0a0"
 
@@ -15,14 +13,6 @@ __version__ = "1.0.0a0"
 METADATA_FILE_NAME: str = "info.json"
 PRIMER_FILE_NAME: str = "primer.bed"
 REFERENCE_FILE_NAME: str = "reference.fasta"
-
-SCHEMES_ARCHIVE_URL = os.environ.get(
-    "PRIMASCHEMA_SCHEMES_ARCHIVE_URL",
-    "https://github.com/pha4ge/primer-schemes/archive/refs/heads/main.tar.gz",
-)
-CACHE_DIR = Path(
-    os.environ.get("PRIMASCHEMA_CACHE_DIR") or user_data_dir("primaschema", "PHA4GE")
-)
 
 PKG_DIR = Path(
     os.environ.get(
